requirements.txt
PyYAML==6.0.2
opencv-python==4.12.0.88
numpy==1.24.4
torch==2.4.1+cu121
ultralytics==8.3.192
pandas==2.0.3
scikit-learn
triton==3.0.0
tritonclient==2.60.0
